+++ forui/lib/src/widgets/radio.dart
@@ -27,7 +27,7 @@ class FRadio extends StatefulWidget {
   /// If the value is present, the radio is in an error state.
   final Widget? error;
 
-  //// {@macro forui.foundation.doc_templates.semanticLabel}
+  //// {@macro forui.foundation.doc_templates.semanticsLabel}
   final String? semanticLabel;
 
   /// The current value of the radio.