+++ forui/test/src/widgets/select_menu_tile/select_menu_tile_test.dart
@@ -15,14 +15,14 @@ void main() {
       await tester.pumpWidget(
         TestScaffold.app(
           child: FSelectMenuTile(
-            groupController: controller,
+            controller: controller,
             prefixIcon: FIcon(FAssets.icons.calendar),
             label: const Text('Label'),
             description: const Text('Description'),
             title: const Text('Repeat'),
             subtitle: const Text('Fee, Fo, Fum'),
             details: const Text('None'),
-            menu: [
+            menuTiles: [
               FSelectTile(title: const Text('Item 1'), value: 1),
               FSelectTile(title: const Text('Item 2'), value: 2),
             ],
@@ -44,14 +44,14 @@ void main() {
       await tester.pumpWidget(
         TestScaffold.app(
           child: FSelectMenuTile(
-            groupController: controller,
+            controller: controller,
             prefixIcon: FIcon(FAssets.icons.calendar),
             label: const Text('Label'),
             description: const Text('Description'),
             title: const Text('Repeat'),
             subtitle: const Text('Fee, Fo, Fum'),
             details: const Text('None'),
-            menu: [
+            menuTiles: [
               FSelectTile(title: const Text('Item 1'), value: 1),
               FSelectTile(title: const Text('Item 2'), value: 2),
             ],
@@ -76,14 +76,14 @@ void main() {
         TestScaffold.app(
           child: FSelectMenuTile(
             popover: const FSelectMenuTilePopoverProperties(autoHide: true),
-            groupController: controller,
+            controller: controller,
             prefixIcon: FIcon(FAssets.icons.calendar),
             label: const Text('Label'),
             description: const Text('Description'),
             title: const Text('Repeat'),
             subtitle: const Text('Fee, Fo, Fum'),
             details: const Text('None'),
-            menu: [
+            menuTiles: [
               FSelectTile(title: const Text('Item 1'), value: 1),
               FSelectTile(title: const Text('Item 2'), value: 2),
             ],