+++ samples/lib/widgets/persistent_sheet.dart
@@ -186,7 +186,7 @@ class _DraggableState extends StatefulSampleState<DraggablePersistentSheetPage> 
                 ),
                 child: FTileGroup.builder(
                   count: 25,
-                  controller: controller,
+                  scrollController: controller,
                   tileBuilder: (context, index) => FTile(title: Text('Tile $index')),
                 ),
               ),